+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 1.0.4 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Copy future object from batch request to request
 
 
 1.0.3 (2018-06-27)